% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemapper-package.R
\docType{package}
\name{basemapper-package}
\alias{basemapper}
\alias{basemapper-package}
\title{basemapper: Headless Spatial Basemap Renderer}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Renders styled spatial basemaps headlessly using a Rust core (maplibre-rs + wgpu) and exposes the result as raw RGBA pixel arrays to R. Integrates with ggplot2 via a lazy GeomBasemap ggproto layer and with tmap v4 via tm_basemap().
}
\author{
\strong{Maintainer}: Pukar Bhandari \email{arpuuuk@gmail.com}

Authors:
\itemize{
  \item Pukar Bhandari \email{arpuuuk@gmail.com}
}

}
