% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_utils.R
\name{.maybe_message_projected}
\alias{.maybe_message_projected}
\title{Emit a message when bbox coordinates fall outside WGS-84 bounds.}
\usage{
.maybe_message_projected(xmin, ymin, xmax, ymax)
}
\description{
Emit a message when bbox coordinates fall outside WGS-84 bounds.
}
\keyword{internal}
