% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_basemap.R
\name{makeContent.BasemapGrob}
\alias{makeContent.BasemapGrob}
\title{makeContent method for BasemapGrob.}
\usage{
\method{makeContent}{BasemapGrob}(x)
}
\description{
Called by grid at draw time (inside an active viewport).  Converts npc
to pixel dimensions, fetches tiles, and returns a rasterGrob child.
}
