% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_basemap.R
\name{GeomBasemap}
\alias{GeomBasemap}
\title{ggproto class for basemap rendering inside a ggplot2 panel.}
\usage{
GeomBasemap
}
\description{
Tile fetching is deferred to draw time so that \code{coord_sf()} has already
established the panel's spatial extent.
}
\keyword{internal}
